fixed unittests in test_vectorizer.py and renamed plural
